+++ lib/src/constants/mp_constants.dart
@@ -234,6 +234,8 @@ const double mpCompassCardinalDirectionsFontSizeFactor = 0.1;
 const double mpCompassCardinalDirectionsTextOffsetFactor = 0.83;
 const double mpCompassCentralCircleFactor = 0.3;
 const double mpSlopeLinePointDefaultLSize = 40.0;
+const int mpLSizeOptionDecimalPlaces = 1;
+const int mpOrientationOptionDecimalPlaces = 1;
 
 const double thCentimeterToMeter = 0.01;
 const double thMeterToCentimeter = 100.0;