+++ CHANGELOG.md
@@ -39,6 +39,7 @@
   * Using lsize and orientation behaviour inspired by xThreion original one.
   * LSize and orientation interactive arrow should appear immediately when starting to set either of them.
   * Reusing Path created to draw compass arrow to also fuel the click test at MPTH2FileEditStateEditSingleLine.
+  * Decimal places for double to string conversion of LSize and Orientation options as constants.
 
 ## 0.2.35 - 2026-02-13 - The [Chiquita Bacana's daughter](https://www.youtube.com/watch?v=QajgvvuFOSg) release
 * New features: